requirements.txt
Flask
psycopg2-binary
Flask-JWT-Extended
Flask-Bcrypt